+++ .gitignore
@@ -4,3 +4,4 @@
 !.vscode/launch.json
 !.vscode/extensions.json
 *.code-workspace
+__pycache__
