+++ organizer.py
@@ -3,10 +3,27 @@ import operator
 from datetime import datetime, timedelta, date
 import re
 import json
+from phonebook import AddressBook, Record
 
 # Dictionary with working days for sort operation
 days = {0: "Monday", 1: "Tuesday", 2: "Wednesday", 3: "Thursday", 4: "Friday"}
 
+HELP = (
+    """load - load data from json file. Default filename - data.bin
+load <filename> - load data from specified json file
+write - write book data into file. Default filename - data.bin
+write <filename> - write data to cpecified json file
+add <contact_name> <phone_number> - Add contact with a phone number. Phone number must be 10 digits
+change <contact_name> <old_phone_number> <new_phone_number> - Change existing phone number for existing contact
+add-birthday <contact_name> <date> - Add birthday data for existing contact or new contact with birthday only. Date format DD.MM.YYYY
+phone <contact_name> - Display phones for contact
+show-birthday <contact_name> - Display birthday data for contact
+birthdays - Show birtdays for next 7 days
+birthdays <date> - Show birtdays for next 7 days from selected date. Date format DD.MM.YYYY
+delete <contact_name> - Delete contact data from book
+hello - get a greeting
+close or exit - exit the program
+""")
 
 # Decorate some access errors
 def input_error(func):
@@ -30,126 +47,6 @@ def parse_input(user_input):
     return cmd, *args
 
 
-class Field:
-    def __init__(self, value):
-        self.value = value
-
-    def __str__(self):
-        return str(self.value)
-
-
-class Birthday(Field):
-    # реалізація класу
-    def __init__(self, Field):
-        self.__value = None
-        self.value = Field
-
-    @property
-    def value(self):
-        return self.__value
-
-    # Verification for current and previous century, 12 months, 31 days
-    @value.setter
-    def value(self, v):
-        if re.search(
-            "^(0[1-9]|[1,2][0-9]|3[0-1])\.(0[1-9]|1[0-2])\.(19\d\d|20\d\d)$", v
-        ):
-            self.__value = v
-        else:
-            raise ValueError
-
-
-class Name(Field):
-    # реалізація класу
-    def __init__(self, Field):
-        self.value = Field
-
-
-class Phone(Field):
-    # реалізація класу
-    def __init__(self, Field):
-        self.__value = None
-        self.value = Field
-
-    @property
-    def value(self):
-        return self.__value
-
-    # Phone length 10 digits
-    @value.setter
-    def value(self, v):
-        if re.search("^\d{10}$", v):
-            self.__value = v
-        else:
-            raise ValueError
-
-
-class Record:
-    def __init__(self, name):
-        self.name = Name(name)
-        self.phones = []
-
-    def add_birthday(self, value):
-        field = Birthday(value)
-        self.birthday = field
-
-    def add_phone(self, value):
-        field = Phone(value)
-        self.phones.append(field)
-
-    def remove_phone(self, value):
-        res = ""
-        for ph in self.phones:
-            if ph.value == value:
-                res = ph
-        if res:
-            self.phones.remove(res)
-            return True
-        else:
-            return False
-
-    def edit_phone(self, old_value, new_value):
-        for ph in self.phones:
-            if ph.value == old_value:
-                ph.value = new_value
-                return True
-        return False
-
-    def find_phone(self, value):
-        for ph in self.phones:
-            if ph.value == value:
-                return ph
-        return None
-
-    def __str__(self):
-        return f"Contact name: {self.name.value}, phones: {'; '.join(p.value for p in self.phones)}, birthday: {self.birthday if 'birthday' in self.__dict__ else 'NA'} "  # if p.value is not None
-
-
-class AddressBook(UserDict):
-    def add_record(self, record):
-        self.data[record.name] = record
-        return record
-
-    # Search before any other operation
-    def find(self, name):
-        for n in self.data.keys():
-            if n.value == name:
-                return self.data[n]
-        return None
-
-    def delete(self, name):
-        res = None
-        for n in self.data.keys():
-            if n.value == name:
-                res = n
-        if res:
-            del self.data[res]
-        return res
-
-    def show_birthday(self, name):
-        rec = self.find(name)
-        if rec:
-            print(f"{rec.name} birthday: {rec.birthday}")
 
 
 # Add contact with phone or add new phone to existing one
@@ -343,6 +240,9 @@ def write_data(args, book):
         json.dump(contacts, fh)
     return "Book written"
 
+@input_error
+def show_help(args, book):
+    return HELP
 
 # Available operations on contacts
 actions = {
@@ -357,6 +257,7 @@ actions = {
     "add-birthday": add_birthday,
     "show-birthday": show_birthday,
     "birthdays": get_birthdays_per_week,
+    "help": show_help,
 }
 
 book = AddressBook()
